[MPP] Allow smart casts for properties from dependsOn modules
Smartcasts for public properties from different module are not
stable because the module with the property in general can be
compiled separately from a module using it. However, if client
module has dependsOn relation with declaring module their simultaneous
compilation is guaranteed which makes this smart cast safe.

KT-42754 Fixed